Fix: maintain identity of expenses, don't track by $index
The bug when removing an expense came from the 'track by $index', which
in turn had been a (misguided) attempt at fixing another bug. The problem
with tracking by $index becomes apparent here, when the list is reduced.
The shift in indices confused the ng-repeater, causing things to look
weird: the wrong entry would be deleted, and under certain circumstances,
other entries would appear to be duplicated. However, upon reloading the
page, it would become apparent that the correct entry had been removed.

The fix is to maintain the identity of the expenses objects in the
bill.expenses collection.