Fix: Share input for new participant would be empty
When adding an expense with a new participant, the input fields for their
number of shares for all other expenses would be added, but would remain
empty, even though the actual sharingModels were all updated correctly.

The reason is that since the expenseController works with a copy of the
expense object, it wouldn't automatically receive updates when an expense
was changed from the outside, which is the case when shares are kept
up to date with the expenses.

Bug was introduced with 3648c362e1f3de, where using the original instance
instead of a copy was forgotten.